Post simulation calculations
Section "Post Simulation" is added in cir_to_ivp.py. It reads out all the matrices used for simulation. Currently, we can plot iB, PhiK, PhiL, and Vz.

The more difficult outputs are the currents through the inductors. When a mutual coupling is present, the current of the two inductors need to be solved using a system of algebraic equations. I think I will need more time to see how to do it generally.

Somethings I may need to do:
1. Reorder the rows and columns L, L_K, L_LK matrices so they match the numbering of inductors in the circuit file. This could slow down the simulation slightly, so I am trying to avoid it.

2. Create a matrix "A" that combines L_t and F_KB. Multiply it with "I"= (IBs I_L1 ... I_K1 ...). It is equal to "Phi" = all the fluxes at each inductor (0 for rows correspond to bias current). This should allow us to solve for current through tree inductors.

3. Using the bias and tree inductor currents and F_KL to solve for currents through the chord inductors.